vg: support for gopherjs
Move the call to os.Stat under a !js build tag to support compilation for gopherjs.
